# Fix: Bug #443 -- better optimize network.py high-speed network