Fix: Bug #443 -- better optimize network.py high-speed network performance.

## Description

This pull request fixes sub-optimal network performance in the network.py network speed tests on high-speed (>10 Gbps) network devices. It makes three changes:

- When the script automatically runs more than two threads, it runs a quick (1-minute) optimization test with 0.5x, 1x, 1.5x, and 2x the computed number of threads to find the optimal number of threads. This is not done when the number of threads is specified on the command line.
- The script limits the maximum per-thread throughput (via the iperf3 -b option) to 1 Mbps over the theoretical optimal amount, based on the interface speed and number of threads.
- The script uses the -A parameter to iperf3 to set the CPU affinity. This requires computing the affinity, using code lifted from the start-iperf3 script in the maas-cert-server package.

The patch has no effect on slow (10 Mbps and slower) network devices, but on faster ones, the thread optimization tests add about four minutes to the run time, and add some output (see below).

<!--

-->

## Resolved issues

Fixes: https://github.com/canonical/checkbox/issues/443
<!--

-->

## Documentation

Neither Checkbox nor the network.py script has changes to their calling conventions or configuration; however, output is expanded a bit, as in this example:

sudo ./network.py test -i ens1 -t iperf --iperf3 --scan-timeout 36 --fail-threshold 80 --cpu-load-fail-threshold 90 --runtime 30 --num_runs 1 --target 10.1.16.15
INFO:root:Testing ens1 against 10.1.16.15
INFO:root:Have successfully pinged 10.1.16.15 on ens1
INFO:root:--------------- Optimizing Number of Threads ---------------
INFO:root:Testing optimization with 5 threads
INFO:root:Found throughput of 39526 with 5 threads
INFO:root:Testing optimization with 10 threads
INFO:root:Found throughput of 85962 with 10 threads
INFO:root:Testing optimization with 15 threads
INFO:root:Found throughput of 75820 with 15 threads
INFO:root:Testing optimization with 20 threads
INFO:root:Found throughput of 98198 with 20 threads
INFO:root:Setting number of threads to 20.
INFO:root:-------------------- Test Run Number 1 --------------------
INFO:root:Using 20 threads.
INFO:root:NUMA node of ens1 is 0....
INFO:root:Connecting to port 5201 on server....
INFO:root:Connecting to port 5202 on server....
INFO:root:Connecting to port 5203 on server....
INFO:root:Connecting to port 5204 on server....
INFO:root:Connecting to port 5205 on server....
INFO:root:Connecting to port 5206 on server....
INFO:root:Connecting to port 5207 on server....
INFO:root:Connecting to port 5208 on server....
INFO:root:Connecting to port 5209 on server....
INFO:root:Connecting to port 5210 on server....
INFO:root:Connecting to port 5211 on server....
INFO:root:Connecting to port 5212 on server....
INFO:root:Connecting to port 5213 on server....
INFO:root:Connecting to port 5214 on server....
INFO:root:Connecting to port 5215 on server....
INFO:root:Connecting to port 5216 on server....
INFO:root:Connecting to port 5217 on server....
INFO:root:Connecting to port 5218 on server....
INFO:root:Connecting to port 5219 on server....
INFO:root:Connecting to port 5220 on server....
INFO:root:Avg Transfer speed: 89510.53333333335 Mb/s
INFO:root:89.51% of theoretical max 100000 Mb/s
INFO:root:Average CPU utilization: 16.0%
INFO:root:

The "Optimizing Number of Threads" section is new, and summarizes the optimization testing. This section is omitted on slow (10 Gbps and slower) tests.

To take advantage of these changes, the iperf3 server may need to run more threads. This will require changes to the Server Certification documentation (the Self-Test Guide).


## Tests

Testing was done by manually running the test using several servers in Needham. Improvements in performance of about 5-10% were observed after these changes were implemented.