@rodwsmith 2023-05-24
I've made Jeff's suggested changes and squashed the results.

Also, I've changed the description (shown near the top of this page, at least to me). I accidentally left the whole thing commented out on the original submission; I've fixed that mistake. Note that the description includes a sample output from running the revised script. The PR includes code to disable logging when doing the test runs, but adds summaries for each test run. This minimizes the added clutter, but there IS some extra output, to help in case there are problems when this type of output might be helpful.

Finally, I've also created a PR for minor changes to the STG so that users know to launch more instances of iperf3 on the Target system:

https://github.com/canonical/certification-docs/pull/17