@baconYao 2023-06-08
> @baconYao , could you please file a bug report? It looks like this should be easy enough to fix, but it would be helpful to have a bug report to reference in the PR. I don't have an RPi 4, but I do have an earlier version, so I can try testing on it.

Hi @rodwsmith, I file an issue #539.